Refactor: Initialize Koin and update MainActivity for navigation
