+++ app/src/main/java/com/jdlstudios/jdldevbox/app/JDLDevBoxApp.kt
@@ -1,0 +1,18 @@
+package com.jdlstudios.jdldevbox.app
+
+import android.app.Application
+import com.jdlstudios.jdldevbox.di.appModule
+import org.koin.android.ext.koin.androidContext
+import org.koin.core.context.startKoin
+
+class JDLDevBoxApp : Application() {
+    override fun onCreate() {
+        super.onCreate()
+
+        startKoin {
+            androidContext(this@JDLDevBoxApp)
+            modules(appModule)
+        }
+
+    }
+}
